modify handlers of api

+++ staff/judith-domenech/full-stack/final-project/logic/src/racket/comment/deleteComment.js
@@ -1,14 +1,19 @@
 const { validators: { validateId } } = require('commons')
-const { models: { Comment } } = require('data')
+const { models: { Comment, User, Racket } } = require('data')
+const { racket } = require('data/src/schemas')
 
 function deleteComment(userId, commentId) {
     validateId(userId, 'user id')
     validateId(commentId, 'comment id')
 
-    return Comment.deleteOne({ _id: commentId, user: userId })
-        .then(result => {
-            if (result.deletedCount === 0) throw new Error (`comment with id ${commentId} and user id ${userId} does not exist`)
-        })
+    return User.findById(userId).lean()
+    .then(user => {
+        if (!user) throw new Error(`user with id ${userId} does not exist`)
+        return Comment.deleteOne({ _id: commentId, user: userId })
+    })
+    .then(result => {
+        if (result.deletedCount === 0) throw new Error(`comment with id ${commentId} and user id ${userId} does not exist`)
+    })
 }
-
+    
 module.exports = deleteComment
